Add Scriptable Waves. Add Wave manager. Add Utilities class
I'm really happy with the current wave system. even though it is very simple, because it is so simple and easy to use, it tickles my brain in a nice way :).

Also added a Utilities class, I figured since I will most likely will need a lot of re-usable code, I will use this class in the future to call from when needing basic method like correctly rotating the sprite towards a target transform.